Refactored StaticScheduler dot output
Refactored the single PDF output function into 
writeGraph and renderGraph functions (which write 
raw dot or render it, respectively).

Extended renderGraph to take a file type argument,
rather than defaulting to the (non-portable) PDF 
backend.

Modified default behavior to write to a dot file 
named for the Transform being compiled. This is 
more portable and easier to use.